Split command-line code into a breezy-cli crate
Move main.rs (the brz binary), command, optparse, help, pycommand and
i18n out of the breezy library into a new breezy-cli crate. breezy-cli
depends on breezy, never the reverse; cmd-py now reaches these modules
via breezy_cli. The bugs help topic moves alongside the help registry.
