+++ crates/cli/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "breezy-cli"
+version = { workspace = true }
+authors = [ "Martin Packman <gzlist@googlemail.com>", "Jelmer Vernooij <jelmer@jelmer.uk>"]
+edition = "2021"
+default-run = "brz"
+license = "GPL-2.0-or-later"
+description = "Command-line interface for Breezy"
+repository = "https://github.com/breezy-team/breezy"
+homepage = "https://www.breezy-vcs.org/"
+documentation = "https://www.breezy-vcs.org/doc/"
+
+[[bin]]
+name = "brz"
+path = "src/main.rs"
+
+[lib]
+
+[features]
+default = ["i18n", "pyo3"]
+i18n = ["dep:gettext"]
+pyo3 = []
+
+[dependencies]
+breezy = { path = "../.." }
+pyo3 = { workspace = true, features = ["abi3"] }
+gettext = { workspace = true, optional = true }
+log = { workspace = true }
+once_cell = "1"
+regex = "1"
+inventory = "0.3"
+lazy_static = "1.4"
+textwrap = ">=0.13"
+url = { workspace = true }
+encoding = "0.2.33"
+patiencediff = "0.2.1"
