+++ crates/cli/src/command.rs
