+++ crates/cli/src/optparse.rs
