+++ crates/cli/src/pycommand.rs
