+++ crates/cmd-py/Cargo.toml
@@ -8,12 +8,13 @@ crate-type = ["cdylib"]
 
 [features]
 default = ["i18n"]
-i18n = ["gettext", "breezy/i18n"]
+i18n = ["gettext", "breezy-cli/i18n"]
 
 [dependencies]
 pyo3 = { workspace = true, features = ["extension-module"]}
 pyo3-filelike = { workspace = true }
 pyo3-log = { workspace = true }
 breezy = { path = "../..", features = ["pyo3"], default-features = false }
+breezy-cli = { path = "../cli", features = ["pyo3"], default-features = false }
 gettext = { workspace = true, optional = true }
 log = { workspace = true, features=["std"]}
